As part of changes to support atomic instructions directly in Valgrind, add support at the IR level for linked loads and store conditionals, a la the lwarx/stwcx etc insns in ppc, but abstractified suitably.
IRExpr_Load gets a new Bool field, indicating whether it's a normal
load or a load-linked (reservation-setting).

IRStmt_Store gets a new IRTemp field.  If this is IRTemp_INVALID, this
is a normal store.  If it's not IRTemp_INVALID, this is a
store-conditional, and the success/failure bit resulting from the
store is written to the new IRTemp.

There are restrictions on alignment of addresses in LL and SC IR
loads.  See libvex_ir.h for details.

Other small IR defn changes:

* IRMBusEvent loses Imbe_BusLock, Imbe_BusUnlock,
  Imbe_SnoopedStoreBegin, Imbe_SnoopedStoreEnd.  These were all
  semantic kludges and can now be removed.

* ppc32 and ppc64 guest state loses the guest_RESVN field.  This was
  part of a kludge to fake up the behaviour of st{w,d}cx. enough to
  make threaded code work.  It is no longer necessary.

The rest of these changes just pushes these through the compilation
pipeline in the normal way.  One minor notable point is that iropt
considers a linked load as inhibiting tree-building, so as to
guarantee that it will not reorder linked-loads w.r.t. any other
loads.



git-svn-id: svn://svn.valgrind.org/vex@1898 8f6e269a-dfd6-0310-a8e1-e2731360e62c