Add logic to allow front ends to speculatively continue adding guest instructions into IRSBs (superblocks) after conditional branches. Currently only the x86 and amd64 front ends support this.  The assumption is that backwards conditional branches are taken and forwards conditional branches are not taken, which is generally regarded as plausible and is particularly effective with code compiled by gcc at -O2, -O3 or -O -freorder-blocks (-freorder-blocks is enabled by default at -O2 and above).
Is disabled by default.  Has been seen to provide notable speedups
(eg, --tool=none for perf/bz2), and reduces the number of
block-to-block transitions dramatically, by up to half, but usually
makes programs run more slowly.  Increases the amount of generated
code by at least 15%-20% and so is a net liability in terms of icache
misses and JIT time.



git-svn-id: svn://svn.valgrind.org/vex@1957 8f6e269a-dfd6-0310-a8e1-e2731360e62c