Back out special handling for opcode 00 (VEX r2189). This was added based on the following analysis at the time: (1) during decoding a sequence of insns we run into a 00 opcode (as that     opcode is sometimes used on purpose to force an abort) (2) #1 only happens when chasing through unconditional gotos (3) the path that was decoded in #1 would not be executed because an earlier     side exit in the super block was taken
But chasing through an unconditional branch should not reach an insn that is
not reached at execution time, because
(a) conditional gotos are supposed to terminate a superblock
(b) side exits that appear in the IR of complex insns will transfer control
    to the very same address (for insns that have implicit loops) and/or to
    the address that immediately follows the current insn (fall through)

Therefore, the special handling of opcode 00 was just fighting the
symptom but not the cause.
Most likely a super block was not correctly terminated.


git-svn-id: svn://svn.valgrind.org/vex@2444 8f6e269a-dfd6-0310-a8e1-e2731360e62c