Add a vg_alignof macro
Note that this macro behaves slightly differently for some types
from the gcc __alignof__ and from the equivalent (new) standardised
alignof. 

vg_alignof macro is needed for the "perm_malloc" callers (next commit)
to determine the alignment of small blocks, but might be useful
for other purposes => placed in libvex_basictypes.h, close to offsetof.



git-svn-id: svn://svn.valgrind.org/vex@2643 8f6e269a-dfd6-0310-a8e1-e2731360e62c