This commit just makes white space changes to the three files in commit r2966 so I can fix the commit message for that commit.  The previous commit message was "msg".  The "msg" was the file with the commit message below.
The first attempt to fix the false positive message "Invalid read of size"
was to change to a V128 read instead of four 32-bit reads.  Unfortunately,
this caused some regression test failures that were not caught before
committing the change.  

This patch implements the V128 read without creating any regression failures.
The issue with the previous fix is that the lvx instruction was used to 
do the V128 fetch.  Unfortunately, that instruction takes the effective 
address masks it to make it 16 byte aligned and then does the fetch.  So,
non-aligned fetches do not work correctly.  The fix in this patch does
two aligned fetches with the lvx instruction, calculates a how to permute
the data from the two loads and then permutes the data so the result in the
vector register is the correct value for an unaligned fetch.

The fix is for bugzilla 339433.


git-svn-id: svn://svn.valgrind.org/vex@2967 8f6e269a-dfd6-0310-a8e1-e2731360e62c